Tile Hit Fix
When a tile is hit on one client it now successfully performs that hit on the other client.

Region added for the photon event handler

Ships are still not sent to the other client. A list, locations, has been added to contain the tile's coordinates and the occupying ship type however no implementation on the receiving end has been added to handle this. Deploying the ships based on the information that can be provided will be complicated.