Corvette working (sort of) patch
Issues:
Not all items in locations are looped through. (PlaceSystemEvent line 41)
Tiles are not updated when a Corvette is added to them meaning the game doesn't acknowledge any there being any ships on the board so the game ends in one shot.
Corvette is the only ship type currently syncing as the other are more complicated due to rotation needing consideration.