Start on advanced mode
Creates a game object in menu scene to hold information regarding whether or not the game is going to load into advanced mode. This object is passed through to the game scene and at the moment used to change whether a cube is active. This cube will be replaced by a power up bar at some point